backend/requirements.txt
watchdog
flask
flask_sqlalchemy
flask_migrate
flask_cors
python-dotenv
openai
azure-cognitiveservices-speech
emoji
waitress
